Algorithm & Time complexity
𝒇𝒊𝒏𝒅 𝒕𝒉𝒆 𝒎𝒆𝒅𝒊𝒂𝒏 𝒊𝒏 𝒌 𝒔𝒊𝒛𝒆𝒅 𝒔𝒖𝒃𝒂𝒓𝒓𝒂𝒚😀 || 𝑶(𝒏 * 𝒌*𝒍𝒐𝒈(𝒌)) 🤓 || 𝒖𝒔𝒊𝒏𝒈 𝒂 𝒉𝒆𝒂𝒑 🌴

The median is the middle value in an ordered integer list. If the size of the list is even, there is no middle value. So the median is the mean of the two middle values.

For examples, if arr = [2,3,4], the median is 3.
For examples, if arr = [1,2,3,4], the median is (2 + 3) / 2 = 2.5.
You are given an integer array nums and an integer k. There is a sliding window of size k which is moving from the very left of the array to the very right. You can only see the k numbers in the window. Each time the sliding window moves right by one position.

Return the median array for each window in the original array. 

𝙆-𝙒𝙞𝙣𝙙𝙤𝙬-𝙈𝙚𝙙𝙞𝙖𝙣(𝘼, 𝙠):
we start with creating a heap array called pq when pq[ i ] contains all the elements from indexes i to i + k in A than in each iteration we find the median in the heap pq[ i ] which takes O(k * log(k)) time and append it into the list 

Time complexity : O(n * k * log( k )):
n sized for loop : O(n)
build a heap in each iteration : O(k * log(k))
finding median of heap : O(k * log(k))

space complexity: O(n * k):
n sized array of k sized heap : O(k)